[AOT Refactor] dispatch compile graph
[ghstack-poisoned]
