+++ torch/_functorch/aot_autograd.py
@@ -5,8 +5,6 @@ from functools import partial, wraps
 from typing import Any, Callable, Dict, List, Optional, Tuple
 from unittest.mock import patch
 
-from torch.fx.experimental.proxy_tensor import make_fx
-
 import torch
 import torch.nn as nn
 import torch.utils._pytree as pytree
@@ -19,7 +17,7 @@ from torch._guards import detect_fake_mode, tracing
 from torch._prims_common import CUDARngStateHelper
 from torch._logging import getArtifactLogger
 from torch._subclasses import FakeTensor, FakeTensorMode
-from torch.fx.experimental.proxy_tensor import is_sym_node
+from torch.fx.experimental.proxy_tensor import is_sym_node, make_fx
 from torch.fx.experimental.symbolic_shapes import (
     ShapeEnv, fx_placeholder_vals
 )
@@ -71,13 +69,9 @@ from ._aot_autograd.input_output_analysis import (
     create_graph_signature,
 )
 
-from ._aot_autograd.traced_function_transforms import (
-    fn_input_mutations_to_outputs,
-    fn_prepped_for_autograd,
-    create_joint,
-    create_functionalized_fn,
-    aot_dispatch_subclass,
+from ._aot_autograd.traced_function_transforms import (  # noqa: F401
     create_functional_call,
+    create_joint_function,
 )
 from ._aot_autograd.runtime_wrappers import (
     create_runtime_wrapper,
@@ -87,6 +81,10 @@ from ._aot_autograd.runtime_wrappers import (
 from ._aot_autograd.alias_runtime_wrappers import (
     aot_wrapper_dedupe,
     aot_wrapper_synthetic_base,
+)
+from ._aot_autograd.dispatch_and_compile_graph import (
+    aot_dispatch_base_graph,
+    aot_dispatch_autograd_graph,
 )
 
 zip = strict_zip
@@ -352,68 +350,7 @@ AOT_COUNTER = itertools.count()
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 
-def create_graph(f, args, *, aot_config: AOTConfig) -> torch.fx.GraphModule:
-    with enable_python_dispatcher():
-        fx_g = make_fx(f, decomposition_table=aot_config.decompositions)(*args)
-
-    return fx_g
-
-
 aot_autograd_decompositions = {}
-
-
-def aot_dispatch_base_graph(
-    flat_fn,
-    flat_args: List[Tensor],
-    aot_config: AOTConfig,
-    *,
-    fw_metadata: ViewAndMutationMeta
-) -> Tuple[Callable, List[Any], Optional[SubclassMeta]]:
-    # aot_dispatch_base requires functionalization, but doesn't need to handle as many cases as the autograd case.
-    # The cases that aot_dispatch_base doesn't need to handle include:
-    # - outputs that are aliases of graph intermediates
-    # - outputs that are aliases of graph inputs
-    # While cases that it does need to handle include:
-    # - input mutations (including when inputs are aliases of each other)
-    # - input metadata mutations
-    fn_to_trace = fn_input_mutations_to_outputs(
-        flat_fn,
-        fw_metadata,
-        keep_data_input_mutations=aot_config.keep_inference_input_mutations,
-    )
-
-    fn_to_trace, updated_flat_args = create_functionalized_fn(
-        fn_to_trace, flat_args, meta=fw_metadata, aot_config=aot_config, trace_joint=False)
-
-    fn_to_trace, updated_flat_args_subclasses_desugared, maybe_subclass_meta = aot_dispatch_subclass(
-        fn_to_trace, updated_flat_args, is_joint_structure=False, meta=fw_metadata, fw_only=flat_fn
-    )
-
-    fw_module = create_graph(
-        fn_to_trace,
-        updated_flat_args_subclasses_desugared,
-        aot_config=aot_config,
-    )
-
-    # As long as we opted to remove input mutations, then
-    # there should be *NO* mutating ops in the graph at this point.
-    copy_count = assert_functional_graph(fw_module.graph, allow_input_mutations=aot_config.keep_inference_input_mutations)
-
-    fw_module.graph.eliminate_dead_code()
-    fw_module.recompile()
-
-    copy_count2 = assert_functional_graph(fw_module.graph, allow_input_mutations=aot_config.keep_inference_input_mutations)
-
-    assert copy_count == copy_count2
-
-    if aot_config.enable_log:
-        aot_graphs_log.info("%s", lazy_format_graph_code("Forward graph", fw_module, aot_config.aot_id))
-
-    # TODO: should factor this into a separate function for export that always only returns just the graph.
-    if aot_config.is_export:
-        assert maybe_subclass_meta is None, "aot_export_module does not support tensor subclass inputs for now."
-        return fw_module
-    return fw_module, list(updated_flat_args_subclasses_desugared), maybe_subclass_meta
 
 def aot_dispatch_base(flat_fn, flat_args: List[Tensor], aot_config: AOTConfig, *, fw_metadata: ViewAndMutationMeta):
     fw_module, updated_flat_args, maybe_subclass_meta = aot_dispatch_base_graph(
@@ -476,61 +413,6 @@ def aot_dispatch_base(flat_fn, flat_args: List[Tensor], aot_config: AOTConfig, *
 
     return compiled_fn
 
-# Has the precondition that there
-# are no duplicate arguments in flat_args (e.g., the same Tensor
-# object never shows up twice.  However, two tensor inputs MAY alias
-# the same storage, so long as they have separate TensorImpls.)
-def aot_dispatch_autograd_graph(flat_fn, flat_args: List[Any], aot_config: AOTConfig, *, fw_metadata: ViewAndMutationMeta):
-    # traced_tangents corresponds to the set of outputs in the traced forward that should get grad_outputs in the traced backward.
-    # It includes outputs of the original forward, *and* any updated inputs due to input mutations.
-    # However, it does *not* include any outputs that are aliases of inputs or intermediates, or any metadata-only input mutations.
-    traced_tangents = pytree.tree_map(
-        lambda x: x.detach().contiguous() if isinstance(x, Tensor) else x,
-        fw_metadata.traced_tangents,
-    )
-
-    joint_inputs = (flat_args, traced_tangents)
-
-    fn_prepared_for_autograd = fn_prepped_for_autograd(
-        flat_fn,
-        fw_metadata,
-    )
-    joint_fn_to_trace = create_joint(fn_prepared_for_autograd, aot_config=aot_config)
-
-    joint_fn_to_trace, updated_joint_inputs = create_functionalized_fn(
-        joint_fn_to_trace,
-        joint_inputs,
-        meta=fw_metadata,
-        aot_config=aot_config,
-        trace_joint=True,
-    )
-
-    subclass_tracing_info = aot_dispatch_subclass(
-        joint_fn_to_trace, updated_joint_inputs, is_joint_structure=True, meta=fw_metadata, fw_only=flat_fn
-    )
-
-    joint_fn_to_trace = subclass_tracing_info.plain_tensor_trace_fn
-    updated_joint_inputs = subclass_tracing_info.plain_tensor_args
-    maybe_subclass_meta = subclass_tracing_info.maybe_subclass_meta
-
-    fx_g = create_graph(joint_fn_to_trace, updated_joint_inputs, aot_config=aot_config)
-
-    # There should be *NO* mutating ops in the graph at this point.
-    assert_functional_graph(fx_g.graph, allow_input_mutations=aot_config.keep_inference_input_mutations)
-
-    # Redundant with the check above, but worth having in case tracing introduced
-    # a fake tensor. Unlikely.
-    # See Note: [Fake Modules and AOTAutograd]
-    torch._dynamo.utils.assert_no_fake_params_or_buffers(fx_g)
-    fx_g.graph.eliminate_dead_code()
-    fx_g.recompile()
-    # TODO: in AOTAutograd, we create metadata like _indices_of_inps_to_detach to detect
-    # when we need to manually detach() some inputs in the forward.
-    # Higher order ops might eventually need to do the same.
-    if aot_config.is_export:
-        assert maybe_subclass_meta is None, "aot_export_module does not support tensor subclass inputs for now."
-        return fx_g
-    return fx_g, updated_joint_inputs, maybe_subclass_meta
 
 def aot_dispatch_autograd(flat_fn, flat_args: List[Any], aot_config: AOTConfig, *, fw_metadata: ViewAndMutationMeta):
     fx_g, joint_inputs, maybe_subclass_meta = aot_dispatch_autograd_graph(flat_fn, flat_args, aot_config, fw_metadata=fw_metadata)
